move and rename plrStudentModel and plrStudent.sql to plr folder in partial + update require statements
