+++ apps/prairielearn/src/pages/partials/plr/plrScoreboard.sql
