+++ apps/prairielearn/src/pages/plrStudent/plrStudent.js
@@ -5,13 +5,16 @@ var ERR = require('async-stacktrace');
 var express = require('express');
 var router = express.Router();
 // Query Stuff
-var sqldb = require('@prairielearn/postgres');
-var sql = sqldb.loadSqlEquiv(__filename);
+// var sqldb = require('@prairielearn/postgres');
+// var path = require('path');
+//
+// var sqlFilePath = path.join(__dirname, '../partials/plr/plrScoreboard.sql');
+// var sql = sqldb.loadSqlEquiv(sqlFilePath);
 // SSE Stuff
 var sseClients = require('../../sseClients');
 // Models
-const { getLiveResults } = require('./plrStudentModel')
-const { getSeasonalResults } = require('./plrStudentModel')
+const { getLiveResults } = require('../partials/plr/plrScoreboardModel')
+const { getSeasonalResults } = require('../partials/plr/plrScoreboardModel')
 // -------
 // ROUTING
 // -------
